Update migec_summary.Rmd
Line 169: We added unique as the factor levels were reading duplicates and causing an error. 
Line 209: We edited that line because the assemble.log.txt file did not contain a column under the name 'READS DROPPED WITHIN MIG'. It contained individual columns for MIG1 and MIG2 for fastq1 and fastq2. We summed it to suit the function. 
We suppressed the warnings - `guides(<scale> = FALSE)` is deprecated. Please use `guides(<scale> = "none")` instead. We DID NOT update the plotting functions, instead we suppressed the warning message to temporarily ignore the warnings on plots.